+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.2.0] - TODO
 
+### Added
+- Add bottom toolbar.
+
 ### Changed
 - Improve hold picker UI.
 - Fix total hold counter being always set to 0.